+++ custom_components/wemportal/switch.py
@@ -3,6 +3,7 @@ Switch platform for wemportal component
 """
 
 import logging
+from typing import Any
 
 from homeassistant.components.switch import SwitchDeviceClass, SwitchEntity
 from homeassistant.config_entries import ConfigEntry
@@ -116,7 +117,7 @@ class WemPortalSwitch(WemPortalEntity, SwitchEntity):
     @property
     def extra_state_attributes(self):
         """Return the state attributes of this device."""
-        attributes = {}
+        attributes: dict[str, Any] = {}
         if self._last_updated is not None:
             attributes["Last Updated"] = self._last_updated
         return attributes